refactor(agent-engine): narrow process file handle
